backend/requirements.txt

fastapi
uvicorn
requests
python-dotenv
chromadb
llama-index
llama-index-readers-github
llama-index-llms-gemini
llama-index-embeddings-gemini
llama-index-vector-stores-chroma
google-generativeai
arq
aiohttp
pydantic-settings
isort
pytest
